backend/requirements.txt
fastapi
uvicorn
pydantichttpx
